Add component state method for checking if current step is between two markers. Use stage 5 as example stage and change incl_range calls.
